requirements.txt
streamlit==1.31.0
torch>=2.2.0
transformers>=4.37.2
sentencepiece>=0.1.99
protobuf>=4.25.2
requests>=2.31.0
accelerate>=0.26.1
sacremoses>=0.1.1
python-dotenv==1.0.0
langdetect==1.0.9